+++ src/api/dto/sendMessage.dto.ts
@@ -70,7 +70,7 @@ export class SendPollDto extends Metadata {
   messageSecret?: Uint8Array;
 }
 
-export type MediaType = 'image' | 'document' | 'video' | 'audio';
+export type MediaType = 'image' | 'document' | 'video' | 'audio' | 'ptv';
 
 export class SendMediaDto extends Metadata {
   mediatype: MediaType;
@@ -82,6 +82,10 @@ export class SendMediaDto extends Metadata {
   media: string;
 }
 
+export class SendPtvDto extends Metadata {
+  video: string;
+}
+
 export class SendStickerDto extends Metadata {
   sticker: string;
 }
@@ -90,15 +94,21 @@ export class SendAudioDto extends Metadata {
   audio: string;
 }
 
-export type TypeButton = 'reply' | 'copy' | 'url' | 'call';
+export type TypeButton = 'reply' | 'copy' | 'url' | 'call' | 'pix';
+
+export type KeyType = 'phone' | 'email' | 'cpf' | 'cnpj' | 'random';
 
 export class Button {
   type: TypeButton;
-  displayText: string;
+  displayText?: string;
   id?: string;
   url?: string;
   copyCode?: string;
   phoneNumber?: string;
+  currency?: string;
+  name?: string;
+  keyType?: KeyType;
+  key?: string;
 }
 
 export class SendButtonsDto extends Metadata {